refactor side menu component import paths
