--- src/components/sideMenu/menuIcons/HistoricoNexusIcon.tsx
@@ -1,56 +1,0 @@
-import React, { ForwardedRef } from 'react';
-
-interface HomeIconProps {
-  [key: string]: any;
-}
-
-const HistoricoNexusIcon = ({ ...props }: HomeIconProps, svgRef: ForwardedRef<SVGSVGElement>) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="28"
-    height="28"
-    viewBox="0 0 31 31"
-    fill="none"
-    ref={svgRef}
-    {...props}
-  >
-    <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M7.03125 4.24997C5.91237 4.24997 4.83931 4.69444 4.04814 5.48561C3.25697 6.27678 2.8125 7.34984 2.8125 8.46872V21.125C2.8125 22.2438 3.25697 23.3169 4.04814 24.1081C4.83931 24.8992 5.91237 25.3437 7.03125 25.3437H11.25C12.0266 25.3437 12.6562 25.9733 12.6562 26.75C12.6562 27.5266 12.0266 28.1562 11.25 28.1562H7.03125C5.16645 28.1562 3.37802 27.4154 2.05941 26.0968C0.74079 24.7782 0 22.9898 0 21.125V8.46872C0 6.60392 0.74079 4.81549 2.05941 3.49688C3.37802 2.17826 5.16645 1.43747 7.03125 1.43747H22.5C24.3648 1.43747 26.1532 2.17826 27.4718 3.49688C28.7905 4.81549 29.5312 6.60392 29.5312 8.46872V12.6875C29.5312 13.4641 28.9016 14.0937 28.125 14.0937C27.3483 14.0937 26.7187 13.4641 26.7187 12.6875V8.46872C26.7187 7.34984 26.2743 6.27678 25.4831 5.48561C24.6919 4.69444 23.6189 4.24997 22.5 4.24997H7.03125Z"
-      fill="#052A76"
-    />
-    <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M9.84376 0.0312195C10.6204 0.0312195 11.25 0.660819 11.25 1.43747V4.24997C11.25 5.02662 10.6204 5.65622 9.84376 5.65622C9.06711 5.65622 8.43751 5.02662 8.43751 4.24997V1.43747C8.43751 0.660819 9.06711 0.0312195 9.84376 0.0312195Z"
-      fill="#052A76"
-    />
-    <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M19.6875 0.0312195C20.4642 0.0312195 21.0938 0.660819 21.0938 1.43747V4.24997C21.0938 5.02662 20.4642 5.65622 19.6875 5.65622C18.9109 5.65622 18.2813 5.02662 18.2813 4.24997V1.43747C18.2813 0.660819 18.9109 0.0312195 19.6875 0.0312195Z"
-      fill="#052A76"
-    />
-    <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M0 9.87503C0 9.09838 0.629599 8.46878 1.40625 8.46878H28.125C28.9016 8.46878 29.5312 9.09838 29.5312 9.87503C29.5312 10.6517 28.9016 11.2813 28.125 11.2813H1.40625C0.629599 11.2813 0 10.6517 0 9.87503Z"
-      fill="#052A76"
-    />
-    <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M25.6038 19.6286C26.1529 20.1778 26.1529 21.0682 25.6038 21.6173L23.4944 23.7267C22.9452 24.2759 22.0548 24.2759 21.5056 23.7267C20.9565 23.1775 20.9565 22.2871 21.5056 21.738L23.615 19.6286C24.1642 19.0794 25.0546 19.0794 25.6038 19.6286Z"
-      fill="#052A76"
-    />
-    <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M22.5 16.9063C19.3934 16.9063 16.875 19.4247 16.875 22.5313C16.875 25.6379 19.3934 28.1563 22.5 28.1563C25.6066 28.1563 28.125 25.6379 28.125 22.5313C28.125 19.4247 25.6066 16.9063 22.5 16.9063ZM14.0625 22.5313C14.0625 17.8714 17.8401 14.0938 22.5 14.0938C27.1599 14.0938 30.9375 17.8714 30.9375 22.5313C30.9375 27.1912 27.1599 30.9688 22.5 30.9688C17.8401 30.9688 14.0625 27.1912 14.0625 22.5313Z"
-      fill="#052A76"
-    />
-  </svg>
-);
-
-export default React.forwardRef(HistoricoNexusIcon);
