--- src/components/sideMenu/menuIcons/ParticiparNexusIcon.tsx
@@ -1,36 +1,0 @@
-import React, { ForwardedRef } from 'react';
-
-interface HomeIconProps {
-  [key: string]: any;
-}
-
-const ParticiparNexusIcon = ({ ...props }: HomeIconProps, svgRef: ForwardedRef<SVGSVGElement>) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="28"
-    height="28"
-    viewBox="0 0 28 28"
-    fill="none"
-    ref={svgRef}
-    {...props}
-  >
-    <path
-      d="M24.4999 0.5C23.323 0.5 21.8269 1.256 21.3839 1.49398L15 0.517016C14.103 0.517016 11.938 0.575 11.236 1.09203L7.236 3.575C7.04602 3.69298 6.95799 3.92202 7.01902 4.13698C7.08104 4.352 7.27702 4.49998 7.5 4.49998H12C12.04 4.49998 12.079 4.49497 12.118 4.48798L12.788 4.32298C13.2209 4.25998 13.6529 4.38898 13.982 4.67398C14.312 4.95898 14.5 5.372 14.5 5.80798C14.5 6.155 14.392 6.48598 14.187 6.767L12.614 8.92798C12.175 9.58597 12.262 10.47 12.822 11.03C13.147 11.354 13.573 11.517 14 11.517C14.427 11.517 14.853 11.354 15.178 11.029L17.713 8.49397C18.66 8.44095 19.526 8.03197 20.168 7.37197C21.056 8.19495 23.898 8.45595 24.461 8.49898C24.474 8.49997 24.487 8.49997 24.5 8.49997C24.625 8.49997 24.747 8.45295 24.839 8.36698C24.942 8.27197 25 8.13898 25 8V1.00002C25 0.724016 24.7759 0.5 24.4999 0.5ZM20.502 3.842L19.925 6.03102C19.451 6.938 18.526 7.50003 17.5 7.50003C17.367 7.50003 17.24 7.55202 17.146 7.64605L14.47 10.322C14.219 10.5731 13.78 10.5731 13.529 10.322C13.305 10.0991 13.27 9.74506 13.434 9.49906L14.996 7.35505C15.326 6.90205 15.5 6.36603 15.5 5.80606C15.5 5.08105 15.186 4.39208 14.638 3.91705C14.185 3.52503 13.612 3.31105 12.996 3.31105C12.8661 3.31105 12.733 3.32103 12.6 3.34006L11.941 3.5H9.25298L11.796 1.91998C12.109 1.68997 13.486 1.517 14.924 1.511L20.855 2.42398C20.758 2.771 20.632 3.26098 20.502 3.842ZM24 7.45002C22.919 7.328 21.591 7.079 21.025 6.764C21.1249 5.792 21.301 4.85 21.475 4.07202L21.933 2.33403C22.395 2.09103 23.256 1.69602 24 1.55305V7.45002Z"
-      fill="#052A76"
-    />
-    <path
-      d="M19.992 18.918L18.992 12.918C18.952 12.677 18.744 12.5 18.499 12.5H16C16 12.224 15.776 12 15.5 12C15.224 12 14.9999 12.224 14.9999 12.5V12.995C14.9999 12.997 14.999 12.998 14.999 13C14.999 13.002 14.9999 13.003 14.9999 13.005V15.5H5.99996V6.5H13C13.276 6.5 13.5 6.27598 13.5 5.99998C13.5 5.72398 13.276 5.49997 13 5.49997H5.49999C5.22399 5.49997 4.99998 5.72398 4.99998 5.99998V12.5H2.5C2.25601 12.5 2.047 12.677 2.00701 12.918L1.00703 18.918C1.00502 18.931 1.01003 18.943 1.00905 18.955C1.00802 18.971 1 18.984 1 19V24C1 24.276 1.22402 24.5 1.50001 24.5H19.5C19.776 24.5 20 24.276 20 24V19C20 18.983 19.992 18.968 19.99 18.952C19.989 18.94 19.994 18.929 19.992 18.918ZM2.923 13.5H4.99998V15.5H4.50001C4.22401 15.5 3.99999 15.724 3.99999 16C3.99999 16.276 4.22401 16.5 4.50001 16.5H16.5C16.776 16.5 17 16.276 17 16C17 15.724 16.776 15.5 16.5 15.5H16V13.5H18.076L18.909 18.5H2.08998L2.923 13.5ZM19 23.5H1.99998V19.5H19V23.5Z"
-      fill="#052A76"
-    />
-    <path
-      d="M24.4999 0.5C23.323 0.5 21.8269 1.256 21.3839 1.49398L15 0.517016C14.103 0.517016 11.938 0.575 11.236 1.09203L7.236 3.575C7.04602 3.69298 6.95799 3.92202 7.01902 4.13698C7.08104 4.352 7.27702 4.49998 7.5 4.49998H12C12.04 4.49998 12.079 4.49497 12.118 4.48798L12.788 4.32298C13.2209 4.25998 13.6529 4.38898 13.982 4.67398C14.312 4.95898 14.5 5.372 14.5 5.80798C14.5 6.155 14.392 6.48598 14.187 6.767L12.614 8.92798C12.175 9.58597 12.262 10.47 12.822 11.03C13.147 11.354 13.573 11.517 14 11.517C14.427 11.517 14.853 11.354 15.178 11.029L17.713 8.49397C18.66 8.44095 19.526 8.03197 20.168 7.37197C21.056 8.19495 23.898 8.45595 24.461 8.49898C24.474 8.49997 24.487 8.49997 24.5 8.49997C24.625 8.49997 24.747 8.45295 24.839 8.36698C24.942 8.27197 25 8.13898 25 8V1.00002C25 0.724016 24.7759 0.5 24.4999 0.5ZM20.502 3.842L19.925 6.03102C19.451 6.938 18.526 7.50003 17.5 7.50003C17.367 7.50003 17.24 7.55202 17.146 7.64605L14.47 10.322C14.219 10.5731 13.78 10.5731 13.529 10.322C13.305 10.0991 13.27 9.74506 13.434 9.49906L14.996 7.35505C15.326 6.90205 15.5 6.36603 15.5 5.80606C15.5 5.08105 15.186 4.39208 14.638 3.91705C14.185 3.52503 13.612 3.31105 12.996 3.31105C12.8661 3.31105 12.733 3.32103 12.6 3.34006L11.941 3.5H9.25298L11.796 1.91998C12.109 1.68997 13.486 1.517 14.924 1.511L20.855 2.42398C20.758 2.771 20.632 3.26098 20.502 3.842ZM24 7.45002C22.919 7.328 21.591 7.079 21.025 6.764C21.1249 5.792 21.301 4.85 21.475 4.07202L21.933 2.33403C22.395 2.09103 23.256 1.69602 24 1.55305V7.45002Z"
-      stroke="#052A76"
-    />
-    <path
-      d="M19.992 18.918L18.992 12.918C18.952 12.677 18.744 12.5 18.499 12.5H16C16 12.224 15.776 12 15.5 12C15.224 12 14.9999 12.224 14.9999 12.5V12.995C14.9999 12.997 14.999 12.998 14.999 13C14.999 13.002 14.9999 13.003 14.9999 13.005V15.5H5.99996V6.5H13C13.276 6.5 13.5 6.27598 13.5 5.99998C13.5 5.72398 13.276 5.49997 13 5.49997H5.49999C5.22399 5.49997 4.99998 5.72398 4.99998 5.99998V12.5H2.5C2.25601 12.5 2.047 12.677 2.00701 12.918L1.00703 18.918C1.00502 18.931 1.01003 18.943 1.00905 18.955C1.00802 18.971 1 18.984 1 19V24C1 24.276 1.22402 24.5 1.50001 24.5H19.5C19.776 24.5 20 24.276 20 24V19C20 18.983 19.992 18.968 19.99 18.952C19.989 18.94 19.994 18.929 19.992 18.918ZM2.923 13.5H4.99998V15.5H4.50001C4.22401 15.5 3.99999 15.724 3.99999 16C3.99999 16.276 4.22401 16.5 4.50001 16.5H16.5C16.776 16.5 17 16.276 17 16C17 15.724 16.776 15.5 16.5 15.5H16V13.5H18.076L18.909 18.5H2.08998L2.923 13.5ZM19 23.5H1.99998V19.5H19V23.5Z"
-      stroke="#052A76"
-    />
-  </svg>
-);
-
-export default React.forwardRef(ParticiparNexusIcon);
